refactor: directory structure #12
