+++ src/main/java/com/gam/api/controller/S3Controller.java
@@ -1,9 +1,8 @@
 package com.gam.api.controller;
 
 import com.gam.api.common.message.ResponseMessage;
-import com.gam.api.dto.s3.request.ImageDeleteRequestDTO;
 import com.gam.api.dto.s3.request.PresignedRequestDTO;
-import com.gam.api.service.S3Service;
+import com.gam.api.service.s3.S3ServiceImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
 import org.springframework.http.ResponseEntity;
@@ -14,7 +13,7 @@ import com.gam.api.common.ApiResponse;
 @RequiredArgsConstructor
 @RequestMapping("api/v1/s3")
 public class S3Controller {
-    private final S3Service s3Service;
+    private final S3ServiceImpl s3Service;
 
     @GetMapping("/image")
     public ResponseEntity<ApiResponse> getPresignedUrl(@RequestParam String fileName) {
@@ -26,11 +25,5 @@ public class S3Controller {
     public ResponseEntity<ApiResponse> getPresignedUrls(@RequestBody PresignedRequestDTO request) {
         val response = s3Service.getPresignedUrls(request);
         return ResponseEntity.ok(ApiResponse.success(ResponseMessage.SUCCESS_GET_PRESIGNED_URLS.getMessage(), response));
-    }
-
-    @DeleteMapping("/image")
-    public ResponseEntity<ApiResponse> getDeletePresignedUrl(@RequestBody ImageDeleteRequestDTO request) {
-        s3Service.getDeletePresignedUrl(request);
-        return ResponseEntity.ok(ApiResponse.success(ResponseMessage.SUCCESS_DELETE_IMAGE.getMessage()));
     }
 }
