+++ src/main/java/com/gam/api/service/social/KakaoSocialService.java
@@ -1,4 +1,4 @@
-package com.gam.api.service;
+package com.gam.api.service.social;
 
 import com.gam.api.config.AuthConfig;
 import com.gam.api.config.jwt.JwtTokenManager;
