+++ src/main/java/com/gam/api/service/social/SocialService.java
@@ -1,4 +1,4 @@
-package com.gam.api.service;
+package com.gam.api.service.social;
 
 import com.gam.api.dto.social.SocialLoginRequestDTO;
 import com.gam.api.dto.social.SocialLoginResponseDTO;
