+++ src/main/java/com/gam/api/service/work/WorkServiceImpl.java
@@ -1,4 +1,4 @@
-package com.gam.api.service;
+package com.gam.api.service.work;
 
 import com.gam.api.common.exception.WorkException;
 import com.gam.api.common.message.ExceptionMessage;
@@ -8,6 +8,7 @@ import com.gam.api.dto.work.response.WorkResponseDTO;
 import com.gam.api.entity.Work;
 import com.gam.api.repository.UserRepository;
 import com.gam.api.repository.WorkRepository;
+import com.gam.api.service.s3.S3ServiceImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
 import org.springframework.http.HttpStatus;
@@ -21,6 +22,7 @@ import javax.persistence.EntityNotFoundException;
 public class WorkServiceImpl implements WorkService {
     private final WorkRepository workRepository;
     private final UserRepository userRepository;
+    private final S3ServiceImpl s3Service;
 
     @Override
     public WorkResponseDTO createWork(Long userId, WorkCreateRequestDTO request) {
@@ -46,12 +48,16 @@ public class WorkServiceImpl implements WorkService {
     public WorkResponseDTO deleteWork(Long userId, WorkDeleteRequestDTO request) {
         val workId = request.workId();
 
-        val work = workRepository.getWorkById(userId)
+        val work = workRepository.getWorkById(workId)
                 .orElseThrow(() -> new EntityNotFoundException(ExceptionMessage.NOT_FOUND_WORK.getMessage()));
 
         if (!work.isOwner(userId)) {
             throw new WorkException(ExceptionMessage.NOT_WORK_OWNER.getMessage(), HttpStatus.BAD_REQUEST);
         }
+
+        val photoUrl = work.getPhotoUrl();
+
+        s3Service.deleteS3Image(photoUrl);
 
         workRepository.deleteById(workId);
